# src/docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "description": "Это API для управления пользователями, товарами и поставщиками.",
        "title": "API документация",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/user/create": {
            "post": {
                "description": "Создает нового пользователя на основе переданных данных",
                "tags": [
                    "users"
                ],
                "summary": "Создание пользователя",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/response.CreateUser"
                            }
                        }
                    },
                    "description": "Данные нового пользователя",
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Пользователь успешно создан",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка в данных",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/delete/{id}": {
            "delete": {
                "description": "Удаляет пользователя по UUID",
                "tags": [
                    "users"
                ],
                "summary": "Удаление пользователя",
                "parameters": [
                    {
                        "description": "UUID пользователя",
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Пользователь успешно удалён",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Неверный формат UUID",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ошибка при удалении пользователя",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/updateAddress/{id}": {
            "put": {
                "description": "Изменяет адрес пользователя по UUID",
                "tags": [
                    "users"
                ],
                "summary": "Обновление адреса пользователя",
                "parameters": [
                    {
                        "description": "UUID пользователя",
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/response.CreateUpdateAddress"
                            }
                        }
                    },
                    "description": "Новый адрес пользователя",
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Адрес успешно изменен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка в параметрах запроса",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/users": {
            "get": {
                "description": "Возвращает список пользователей, отфильтрованных по имени и фамилии",
                "tags": [
                    "users"
                ],
                "summary": "Получение списка пользователей по имени и фамилии",
                "parameters": [
                    {
                        "description": "Имя пользователя",
                        "name": "name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Фамилия пользователя",
                        "name": "surname",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список пользователей",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "array",
                                        "items": {
                                            "$ref": "#/components/schemas/response.UserResponse"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка в параметрах запроса",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ошибка сервера",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/usersList": {
            "get": {
                "description": "Возвращает список пользователей с возможностью пагинации",
                "tags": [
                    "users"
                ],
                "summary": "Получение списка пользователей с пагинацией",
                "parameters": [
                    {
                        "description": "Количество пользователей на странице (по умолчанию 20)",
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "Смещение (по умолчанию 0)",
                        "name": "offset",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список пользователей и флаг has_more",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": true
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка в параметрах запроса",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ошибка сервера",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "servers": [
        {
            "url": "//localhost:5000/api/v1"
        }
    ],
    "components": {
        "schemas": {
...
            "response.CreateUpdateAddress": {
                "type": "object",
                "properties": {
                    "city": {
                        "type": "string"
                    },
                    "country": {
                        "type": "string"
                    },
                    "street": {
                        "type": "string"
                    }
                }
            },
            "response.CreateUser": {
                "type": "object",
                "properties": {
                    "address": {
                        "$ref": "#/components/schemas/response.CreateUpdateAddress"
                    },
                    "birthday": {
                        "type": "string"
                    },
                    "gender": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "surname": {
                        "type": "string"
                    }
                }
            },
...
            "response.UserResponse": {
                "type": "object",
                "properties": {
                    "address": {
                        "type": "string"
                    },
                    "birthday": {
                        "type": "string"
                    },
                    "gender": {
                        "type": "string"
                    },
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "registration_date": {
                        "type": "string"
                    },
                    "surname": {
                        "type": "string"
                    }
                }
            }
        }
    }
}